requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.0
bio==1.6.0
biopython==1.81
biothings-client==0.3.1
black==23.1.0
blinker==1.6.3
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.3
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
distro==1.8.0
executing==1.2.0
gitdb==4.0.10
GitPython==3.1.37
gprofiler-official==1.0.0
graphviz==0.20.1
h11==0.14.0
html-to-json==2.0.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.1.0
jupyter_core==5.3.0
load-dotenv==0.1.0
lxml==4.9.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mygene==3.2.2
mypy==1.1.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.26.0
openai==1.3.6
packaging==23.0
pandas==2.1.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.1.1
plotly==5.17.0
pooch==1.8.0
prompt-toolkit==3.0.38
protobuf==4.24.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.14.0
PyJSONViewer==1.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.0.2
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
streamlit==1.27.2
streamlit-chat==0.1.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
unipressed==1.2.0
urllib3==2.0.6
validators==0.22.0
wcwidth==0.2.6
xlrd==2.0.1
xmltodict==0.12.0
xmltojson==2.0.2
zipp==3.17.0
